pass bedrock_invoke_provider to make_call
