+++ litellm/llms/bedrock/chat/invoke_transformations/base_invoke_transformation.py
@@ -440,6 +440,7 @@ class AmazonInvokeConfig(BaseConfig, BaseAWSLLM):
                 messages=messages,
                 logging_obj=logging_obj,
                 fake_stream=True if "ai21" in api_base else False,
+                bedrock_invoke_provider=self.get_bedrock_invoke_provider(model),
             ),
             model=model,
             custom_llm_provider="bedrock",
@@ -473,6 +474,7 @@ class AmazonInvokeConfig(BaseConfig, BaseAWSLLM):
                 messages=messages,
                 logging_obj=logging_obj,
                 fake_stream=True if "ai21" in api_base else False,
+                bedrock_invoke_provider=self.get_bedrock_invoke_provider(model),
             ),
             model=model,
             custom_llm_provider="bedrock",
